renaming functions, bugFixes, optimisation

+++ src/index.coffee
@@ -29,20 +29,22 @@ class Fingerprint
       # Environment to make hash on files
       environments: ['production']
       # Force fingerprint-brunch to run in all environments when true.
-      alwaysRun: false
+      alwaysRun: true
       # autoReplaceAndHash assets in css/js
       autoReplaceAndHash: true
       # Image pattern format
       # authorized chars : ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~:/?#[]@!$&'()*+,;=
       imagePatterns: new RegExp(/url\([\'\"]?[a-zA-Z0-9\-\/_.:]+\.(woff|woff2|eot|ttf|otf|jpg|jpeg|png|bmp|gif|svg)\??\#?[a-zA-Z0-9\-\/_]*[\'\"]?\)/g)
     }
+    # Map of assets
+    @map = {}
+
     # Merge config
     cfg = @config.plugins?.fingerprint ? {}
     @options[k] = cfg[k] for k of cfg
 
+  # Main method
   onCompile: (generatedFiles) ->
-    map          = {}
-
     # Open files
     for file in generatedFiles
       # Set var with generatedFile
@@ -50,30 +52,55 @@ class Fingerprint
       dir   = path.dirname(filePath)
       ext   = path.extname(filePath)
       base  = path.basename(filePath, ext)
+
       # Search and destory old files if option is enable
       if @options.autoClearOldFiles
         @_clearOldFiles(dir, base, ext)
-      if @options.targets == '*' or (base + ext) in @options.targets
-        fileNewName = filePath
-        if @options.autoReplaceAndHash and fs.existsSync(filePath)
-          @_autoReplaceAndHash(filePath)
-        else if (@config.env[0] in @options.environments) or @options.alwaysRun
-          fileNewName = @_renameFileToHash(filePath)
-        # Unixify & Remove part from original path
-        keyPath = unixify(filePath)
-        keyPath = keyPath.replace @options.srcBasePath, ""
-        realPath = unixify(fileNewName)
-        realPath = realPath.replace @options.destBasePath, ""
-        # Make array for manifest
-        map[unixify(keyPath)] = unixify(realPath)
-    # Merge array to keep not watched files
-    if fs.existsSync @options.manifest
-      @_mergeManifest(map)
-    else
-      @_whriteManifest(map)
 
-  # _autoReplaceAndHash
-  _autoReplaceAndHash: (filePath) ->
+      # Hash only file in targets option key
+      if @options.targets == '*' or (base + ext) in @options.targets
+        if @options.autoReplaceAndHash and fs.existsSync(filePath)
+          # Fingerprint sub files
+          @_findAndReplaceSubAssets(filePath)
+        else
+          @_makeCoffee(filePath)
+
+    # Make array for manifest
+    if fs.existsSync @options.manifest
+      # Merge array to keep not watched files
+      @_mergeManifest()
+    else
+      @_writeManifest()
+
+  _makeCoffee: (filePath) ->
+    fileNewName = filePath
+    if @_isFingerprintable()
+      # Just fingerprint targets
+      fileNewName = @_fingerprintFile(filePath)
+    @_addToMap(filePath, fileNewName)
+
+  # Unixify & Remove part from original path
+  _addToMap: (fileInput, fileOutput) ->
+    fileInput = @_removePathBeforePublic(fileInput)
+    fileOutput = @_removePathBeforePublic(fileOutput)
+
+    # Remove srcBasePath/destBasePath
+    fileInput = fileInput.replace @options.srcBasePath, ""
+    fileOutput = fileOutput.replace @options.destBasePath, ""
+
+    # Adding to @map var
+    @map[fileInput] = fileOutput
+
+  # Remove path before the public
+  _removePathBeforePublic: (path) ->
+    path = unixify path
+    pathPublicIndex = path.indexOf(unixify(@config.paths.public))
+    if (pathPublicIndex != 0)
+      path = path.substring(pathPublicIndex)
+    return path
+
+  # Find dependencied like image, fonts.. Hash them and rewrite files (CSS only for now)
+  _findAndReplaceSubAssets: (filePath) ->
     # read file and match url(**)
     config = @config
     options = @options
@@ -82,7 +109,7 @@ class Fingerprint
     contents = fs.readFileSync(filePath).toString()
     paths = contents.match(@options.imagePatterns)
     if paths != null
-      map = {}
+      # find file into generatedFiles
       Object.keys(paths).forEach (key) ->
         # get path
         match = paths[key]
@@ -95,29 +122,34 @@ class Fingerprint
         paths[key] = paths[key].replace(/(\?|\#)[a-zA-Z0-9\-\/\#_]*/g, '')
         # target exists ?
         targetPath = unixify(path.join(config.paths.public, paths[key]))
-        if fs.existsSync(map[targetPath] || targetPath)
-          if typeof(map[targetPath]) == 'undefined'
+        if fs.existsSync(that.map[targetPath] || targetPath)
+          if typeof(that.map[targetPath]) == 'undefined'
             # rename file
-            targetNewName = that._renameFileToHash(targetPath)
-            map[unixify(targetPath)] = unixify(path.join(config.paths.public, targetNewName.substring(config.paths.public.length)))
+            targetNewName = that._fingerprintFile(targetPath)
+            that._addToMap(targetPath, path.join(config.paths.public, targetNewName.substring(config.paths.public.length)))
           else
-            targetNewName = map[targetPath]
+            targetNewName = that.map[targetPath]
           match = new RegExp(match.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, "\\$&"), 'g')
           # add to map
           # rename path in css
           contents = contents.replace(match, "url('" + unixify(targetNewName.substring(config.paths.public.length)) + finalHash + "')")
+      # END find file into generatedFiles
 
-      console.log map
-      if (config.env[0] in options.environments) or options.alwaysRun
-        filePath = that._generateFileNameHashed(filePath, contents)
+      modifiedFilePath = filePath
+      if @_isFingerprintable()
+        modifiedFilePath = @_fingerprintCompose(filePath, contents)
 
       # write file to generate
-      fs.writeFileSync(filePath, contents, 'utf8')
+      fs.writeFileSync(modifiedFilePath, contents, 'utf8')
+      @_addToMap(filePath, modifiedFilePath)
+    else
+      @_makeCoffee(filePath)
 
-  # Clear all old files
-  # dir
-  # base
-  # ext
+  # IsFingerprintable
+  _isFingerprintable: ->
+    return (@config.env[0] in @options.environments) or @options.alwaysRun
+
+  # Clear all the fingerprinted files
   _clearOldFiles: (dir, base, ext) ->
     # Find and remove file in dir/base-{hash}.ext
     pattern = new RegExp(base + '\\-\\w+\\' + ext + '$');
@@ -126,24 +158,20 @@ class Fingerprint
       filePath = path.normalize(dir + '/' + oldFile)
       if pattern.test oldFile then fs.unlinkSync filePath
 
-  # Generate hash with data of file
-  # filePath
-  _generateHash: (data) ->
+  # Make hash from data of file
+  _makeFingerprintWithData: (data) ->
     shasum = crypto.createHash 'sha1'
     shasum.update(data.toString('utf8'), 'utf8')
     return shasum.digest('hex')[0..@options.hashLength-1]
 
-  # Make the hash
-  # dir
-  # base
-  # ext
-  _generateFileNameHashed: (filePath) ->
+  # Compose file name
+  _fingerprintCompose: (filePath) ->
     hash = null
     if (arguments[1])
-      hash = @_generateHash(arguments[1])
+      hash = @_makeFingerprintWithData(arguments[1])
     else
       data = fs.readFileSync filePath
-      hash = @_generateHash(data)
+      hash = @_makeFingerprintWithData(data)
 
     dir   = path.dirname(filePath)
     ext   = path.extname(filePath)
@@ -151,25 +179,27 @@ class Fingerprint
     newName = "#{base}-#{hash}#{ext}"
     return path.join(dir, newName)
 
-  _renameFileToHash: (filePath) ->
-    fileNewName = @_generateFileNameHashed(filePath)
+  # Rename file with his new fingerprint
+  _fingerprintFile: (filePath) ->
+    fileNewName = @_fingerprintCompose(filePath)
     # Rename file, with hash
-    console.log fileNewName
     fs.renameSync(filePath, fileNewName)
     fileNewName
 
-  # Manifest
+  # Remove existing manifest
   _removeManifest: ->
-    fs.unlinkSync @options.manifest
+    if fs.existsSync(@options.manifest) then fs.unlinkSync @options.manifest
 
-  _whriteManifest: (map) ->
-    output = JSON.stringify map, null, "  "
+  # Write a new manifest
+  _writeManifest: ->
+    output = JSON.stringify @map, null, "  "
     fs.writeFileSync(@options.manifest, output)
 
-  _mergeManifest: (map) ->
+  # Merging existing manifest with new entree
+  _mergeManifest: ->
     manifest = fs.readFileSync @options.manifest, 'utf8'
     manifest = JSON.parse manifest
-    manifest[k] = map[k] for k of map
-    @_whriteManifest(manifest)
+    manifest[k] = @map[k] for k of @map
+    @_writeManifest(manifest)
 
 module.exports = Fingerprint
